+++ src/hooks/useLocalization.ts
@@ -2,7 +2,7 @@ const useLocalization = (city: string) : string => {
     if(city === 'Jakarta') {
         return 'ID';
     } else if(city === 'London') {
-        return 'UK';
+        return 'GB';
     }
     else {
         return 'ID';